refactor(range): move range calculation out to separate function
